start refactoring
