+++ designer/client/components/componentCreate/index.ts
@@ -1,0 +1,1 @@
+export { ComponentCreate } from "./ComponentCreate";
