+++ designer/client/page.js
@@ -7,7 +7,7 @@ import {
 import Flyout from "./flyout";
 import PageEdit from "./page-edit";
 import { Component } from "./component";
-import ComponentCreate from "./component-create";
+import { ComponentCreate } from "./components/componentCreate";
 import { ComponentTypes, clone } from "@xgovformbuilder/model";
 import { withI18n } from "./i18n";
